backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
Pillow
djoser
djangorestframework-simplejwt
